Move source files to lib/ directory and update imports
All source files have been moved under the lib/ directory to better organize the codebase. All internal and README import paths have been updated to reference the new locations under lib/. Minor string changes were made to clarify challenge types in logs and errors.
