+++ lib/stealth/stealth.go
